@@ -13,24 +13,24 @@
     "private": true,
     "dependencies": {
         "@angular-material-extensions/select-country": "^7.0.1",
-        "@angular/animations": "~14.2.1",
-        "@angular/cdk": "^14.2.1",
-        "@angular/common": "~14.2.1",
-        "@angular/compiler": "~14.2.1",
-        "@angular/core": "~14.2.1",
-        "@angular/forms": "~14.2.1",
-        "@angular/material": "^14.2.1",
-        "@angular/material-moment-adapter": "^14.2.1",
-        "@angular/platform-browser": "~14.2.1",
-        "@angular/platform-browser-dynamic": "~14.2.1",
-        "@angular/router": "~14.2.1",
+        "@angular/animations": "~14.2.4",
+        "@angular/cdk": "^14.2.3",
+        "@angular/common": "~14.2.4",
+        "@angular/compiler": "~14.2.4",
+        "@angular/core": "~14.2.4",
+        "@angular/forms": "~14.2.4",
+        "@angular/material": "^14.2.3",
+        "@angular/material-moment-adapter": "^14.2.3",
+        "@angular/platform-browser": "~14.2.4",
+        "@angular/platform-browser-dynamic": "~14.2.4",
+        "@angular/router": "~14.2.4",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@fortawesome/fontawesome-free": "^6.2.0",
         "@ngx-translate/core": "^14.0.0",
         "@ngx-translate/http-loader": "^7.0.0",
-        "@schematics/angular": "^14.2.2",
+        "@schematics/angular": "^14.2.4",
         "@swimlane/ngx-charts": "^20.1.0",
-        "@syncfusion/ej2": "^20.2.48",
+        "@syncfusion/ej2": "^20.3.47",
         "jquery": "^3.6.1",
         "moment": "^2.29.4",
         "ng-click-outside2": "^11.1.0",
@@ -38,35 +38,35 @@
         "ng2-datepicker": "^12.0.0",
         "ngx-file-drag-drop": "^6.0.0",
         "ngx-mat-select-search": "^4.2.1",
-        "remove-accents": "^0.4.2",
-        "rxjs": "~7.5.6",
+        "remove-accents": "^0.4.3",
+        "rxjs": "~7.5.7",
         "svg-country-flags": "^1.2.10"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.2.2",
-        "@angular-devkit/build-webpack": "^0.1402.2",
-        "@angular-eslint/builder": "14.0.4",
-        "@angular-eslint/eslint-plugin": "14.0.4",
-        "@angular-eslint/eslint-plugin-template": "14.0.4",
-        "@angular-eslint/schematics": "14.0.4",
-        "@angular-eslint/template-parser": "14.0.4",
-        "@angular/cli": "~14.2.2",
-        "@angular/compiler-cli": "~14.2.1",
+        "@angular-devkit/build-angular": "^14.2.4",
+        "@angular-devkit/build-webpack": "^0.1402.4",
+        "@angular-eslint/builder": "14.1.2",
+        "@angular-eslint/eslint-plugin": "14.1.2",
+        "@angular-eslint/eslint-plugin-template": "14.1.2",
+        "@angular-eslint/schematics": "14.1.2",
+        "@angular-eslint/template-parser": "14.1.2",
+        "@angular/cli": "~14.2.4",
+        "@angular/compiler-cli": "~14.2.4",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
         "@types/d3-scale": "^4.0.2",
         "@types/d3-selection": "^3.0.3",
-        "@types/node": "^18.7.17",
-        "@typescript-eslint/eslint-plugin": "5.37.0",
-        "@typescript-eslint/parser": "5.37.0",
-        "autoprefixer": "^10.4.9",
-        "eslint": "^8.23.1",
+        "@types/node": "^18.8.1",
+        "@typescript-eslint/eslint-plugin": "5.39.0",
+        "@typescript-eslint/parser": "5.39.0",
+        "autoprefixer": "^10.4.12",
+        "eslint": "^8.24.0",
         "eslint-plugin-import": "latest",
         "eslint-plugin-jsdoc": "latest",
         "eslint-plugin-prefer-arrow": "latest",
-        "postcss": "^8.4.16",
+        "postcss": "^8.4.17",
         "tailwindcss": "^3.1.8",
         "ts-node": "^10.9.1",
-        "typescript": "~4.8.3",
+        "typescript": "~4.8.4",
         "zone.js": "~0.11.8"
     }
 }
